requirements.txt
numpy
pandas
scikit-learn
seaborn
matplotlib
missingno
streamlit
joblib
google-generativeai
python-dotenv
streamlit-mic-recorder
gTTS
